Transition to doubles everywhere

+++ src/geom.rs
@@ -3,7 +3,7 @@ use crate::math::{Ray, Vec3};
 pub struct Intersection {
     pub point: Vec3,
     pub normal: Vec3,
-    pub dist: f32,
+    pub dist: f64,
 }
 
 pub trait Geom {
